+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 6.4
 - New ROSARIO_DISABLE_ADDON_UPLOAD optional config constant in INSTALL.md, INSTALL_es.md & INSTALL_fr.md
 - Add-on zip upload in Modules.inc.php & Plugins.inc.php
 - Translate "Upload" & "Cannot open file." to French & Spanish in rosariosis.po
+- JS fix passwordStrength check when password empty in jquery-passwordstrength.js
 
 Changes in 6.3
 --------------