+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 6.4
 - Translate "Upload" & "Cannot open file." to French & Spanish in rosariosis.po
 - JS fix passwordStrength check when password empty in jquery-passwordstrength.js
 - HTML try to fix autocomplete Username & Password inputs (IE, Edge, Chrome) in Inputs.php, General_Info.inc.php
+- HTML add empty value param to TextInput() in Inputs.php
 
 Changes in 6.3
 --------------